xib学习
1.学习XIB；
2.实践xib在cell中的使用；
3.多种 cell的创建及其实现；
4.约束还是没有解决，星期一再来学习。

+++ 掌上联盟/DT_BaseCell.m
@@ -65,7 +65,7 @@
         _readNum = [UILabel new];
         _readNum.frame = CGRectMake(_imgV.marginX +10, _detilLab.marginY, APPWidth -_imgV.marginY-10, 20);
         NSInteger aint = 10;
-        _readNum.text = [NSString stringWithFormat:@"%ld阅",aint];
+        _readNum.text = [NSString stringWithFormat:@"%ld阅",(long)aint];
         _readNum.textColor = APPColor;
         _readNum.font = [UIFont systemFontOfSize:11];
 //        _readNum.backgroundColor = [UIColor redColor];
